MODEXPS-4 Added FolioContext initialization on tenant register step Scheduler fixed after context changes.
